requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
art==5.2
asgiref==3.4.1
astroid==2.6.5
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
bokeh==2.3.3
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
entrypoints==0.3
filelock==3.0.12
googletrans==3.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.2
inflection==0.5.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.2
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-bokeh==3.0.2
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libretranslatepy==2.1.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.9.5
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
rfc3986==1.5.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.2
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
translate==3.6.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
wrapt==1.12.1
